+++ e2e/sinon/nestjs/automock-sinon-ctor-injection.e2e.test.ts
@@ -1,7 +1,7 @@
 import 'reflect-metadata';
 
-import { UnitReference } from '@automock/core';
-import { TestBed } from '@automock/sinon';
+import { UnitReference } from '@suites/core';
+import { TestBed } from '@suites/sinon';
 import {
   ClassThatIsNotInjected,
   Foo,
@@ -23,7 +23,7 @@ import * as chaiAsPromised from 'chai-as-promised';
 import sinon = require('sinon');
 chai.use(chaiAsPromised);
 
-describe('Automock Sinon / NestJS E2E Test Ctor', () => {
+describe('Suites Sinon / NestJS E2E Test Ctor', () => {
   let unit: NestJSTestClass;
   let unitRef: UnitReference;