+++ packages/adapters/nestjs/src/class-props-reflector.ts
@@ -1,9 +1,5 @@
-import {
-  ClassInjectable,
-  InjectableReflectedType,
-  UndefinedDependencyError,
-} from '@automock/common';
-import { Type } from '@automock/types';
+import { ClassInjectable, InjectableReflectedType, UndefinedDependencyError } from '@suites/common';
+import { Type } from '@suites/types';
 import { PROPERTY_DEPS_METADATA } from '@nestjs/common/constants';
 import { MetadataReflector, ReflectedProperty } from './types';
 import { PropertyReflectionStrategy } from './property-reflection-strategies.static';