+++ packages/core/src/normalize-identifier.static.ts
@@ -1,4 +1,4 @@
-import { InjectableIdentifier } from '@automock/common';
+import { InjectableIdentifier } from '@suites/common';
 
 export function normalizeIdentifier(
   identifier: InjectableIdentifier,