+++ packages/testbeds/sinon/jest.config.ts
@@ -1,8 +1,9 @@
 import { Config } from '@jest/types';
 import baseConfig from '../../../jest.base.config';
+import * as process from 'process';
 
 const config: Partial<Config.InitialOptions> = {
-  ...baseConfig,
+  ...baseConfig(process.env.COVERAGE_DIR),
   id: 'sinon',
   displayName: 'sinon',
 };